+++ src/containers/navigation-icons/NavigationIcons.styles.js
